requirements.txt
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.49.1
